Refactor calling database in AuthController
This commit will fix the issue if we call $this->table. Perhaps when we
store table caller to properties it will persist and eventually the
following call will fail. Thus, we need to return a new instance of
table call so the caller will get a fresh instance of table.

Rise $this->table(), a private method that can be call if we want a new
instance of users table, replacing \Builder::table('users'). A much
welcoming existence to our codebase.